+++ RELEASENOTES.md
@@ -4,3 +4,5 @@
 - Peepholes now accept multiple layers as `target_layers` argument
 - Images, Labels, Results, and Outputs were moved to the `corevector._actds` instead of `corevector._cvsds`
 - `get_corevectors()` now accepts a generic dimensionality reduction function
+- `ClassifierBase` receives a `CoreVectors` class as argument instead of dataloader
+- `ClassifierBase` can be saved and loaded (error saying it is not fitted when loading). It receives `path` and `name` arguments for saving.